requirements.txt
yfinance~=0.1.63
pandas~=1.3.2
alpaca-trade-api~=1.3.0
requests~=2.26.0
pushbullet.py~=0.12.0
ta~=0.8.0
numpy~=1.20.3